refactoring das fixtures
