+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/Common/CategoryUseCasesBaseFixture.cs
@@ -1,0 +1,46 @@
+﻿using FC.CodeFlix.Catalog.Application.Interfaces;
+using FC.CodeFlix.Catalog.Domain.Entity;
+using FC.CodeFlix.Catalog.Domain.Repository;
+using FC.CodeFlix.Catalog.UnitTests.Common;
+using Moq;
+using System;
+
+namespace FC.CodeFlix.Catalog.UnitTests.Application.Common;
+
+public abstract class CategoryUseCasesBaseFixture: BaseFixture
+{
+    public Mock<ICategoryRepository> GetRepositoryMock() => new();
+
+    public Mock<IUnitOfWork> GetUnitOfWorkMock() => new();
+
+    public string GetValidCategoryName()
+    {
+        var categoryName = "";
+
+        while (categoryName.Length < 3)
+            categoryName = Faker.Commerce.Categories(1)[0];
+
+        if (categoryName.Length > 255)
+            categoryName = categoryName[..255];
+
+        return categoryName;
+    }
+
+    public string GetValidCategoryDescription()
+    {
+        var categoryDescription = Faker.Commerce.ProductDescription();
+
+
+        if (categoryDescription.Length > 10_000)
+            categoryDescription = categoryDescription[..10_000];
+
+
+        return categoryDescription;
+    }
+
+    public bool GetRandomBoolean()
+        => (new Random().NextDouble()) < 0.5;
+
+    public Category GetExampleCategory()
+        => new(GetValidCategoryName(), GetValidCategoryDescription(),GetRandomBoolean());
+}
