+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/CreateCategory/CreateCategoryTest.cs
@@ -4,7 +4,6 @@ using FC.CodeFlix.Catalog.Domain.Exceptions;
 using FluentAssertions;
 using Moq;
 using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
