+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/CreateCategory/CreateCategoryTestFixture.cs
@@ -1,48 +1,16 @@
-﻿using FC.CodeFlix.Catalog.Application.Interfaces;
-using FC.CodeFlix.Catalog.Application.UseCases.Category.CreateCategory;
-using FC.CodeFlix.Catalog.Domain.Repository;
-using FC.CodeFlix.Catalog.UnitTests.Common;
-using Moq;
-using System;
+﻿using FC.CodeFlix.Catalog.Application.UseCases.Category.CreateCategory;
+using FC.CodeFlix.Catalog.UnitTests.Application.Common;
 using Xunit;
 
 namespace FC.CodeFlix.Catalog.UnitTests.Application.CreateCategory;
 
 [CollectionDefinition(nameof(CreateCategoryTestFixture))]
-public class CreateCategoryTestFixtureCollection: ICollectionFixture<CreateCategoryTestFixture>
+public class CreateCategoryTestFixtureCollection : ICollectionFixture<CreateCategoryTestFixture>
 {
 
 }
-public class CreateCategoryTestFixture : BaseFixture
+public class CreateCategoryTestFixture : CategoryUseCasesBaseFixture
 {
-    public string GetValidCategoryName()
-    {
-        var categoryName = "";
-
-        while (categoryName.Length < 3)
-            categoryName = Faker.Commerce.Categories(1)[0];
-
-        if (categoryName.Length > 255)
-            categoryName = categoryName[..255];
-
-        return categoryName;
-    }
-
-    public string GetValidCategoryDescription()
-    {
-        var categoryDescription = Faker.Commerce.ProductDescription();
-
-
-        if (categoryDescription.Length > 10_000)
-            categoryDescription = categoryDescription[..10_000];
-
-
-        return categoryDescription;
-    }
-
-    public bool GetRandomBoolean()
-        =>  (new Random().NextDouble()) < 0.5;
-
     public CreateCategoryInput GetInput()
         => new(GetValidCategoryName(), GetValidCategoryDescription(), GetRandomBoolean());
 
@@ -51,7 +19,7 @@ public class CreateCategoryTestFixture : BaseFixture
         var invalidInputShortName = GetInput();
         invalidInputShortName.Name = invalidInputShortName.Name.Substring(0, 2);
         return invalidInputShortName;
-        
+
     }
 
     public CreateCategoryInput GetInvalidInputTooLongName()
@@ -81,9 +49,4 @@ public class CreateCategoryTestFixture : BaseFixture
     }
 
 
-    public Mock<ICategoryRepository> GetRepositoryMock() => new();
-
-    public Mock<IUnitOfWork> GetUnitOfWorkMock() => new();
-
-    
 }
