+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/DeleteCategory/DeleteCategoryTest.cs
@@ -23,7 +23,7 @@ public class DeleteCategoryTest
     {
         var repositoryMock = _fixture.GetRepositoryMock();
         var unitOfWork = _fixture.GetUnitOfWorkMock();
-        var categoryExample = _fixture.GetValidCategory();
+        var categoryExample = _fixture.GetExampleCategory();
 
         repositoryMock.Setup(x => x.Get(categoryExample.Id, It.IsAny<CancellationToken>()
                 )).ReturnsAsync(categoryExample);
