+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/ListCategories/ListCategoriesTestFixture.cs
@@ -1,9 +1,7 @@
 ﻿using FC.CodeFlix.Catalog.Application.UseCases.Category.ListCategories;
 using FC.CodeFlix.Catalog.Domain.Entity;
-using FC.CodeFlix.Catalog.Domain.Repository;
 using FC.CodeFlix.Catalog.Domain.SeedWork.SearchableRepository;
-using FC.CodeFlix.Catalog.UnitTests.Common;
-using Moq;
+using FC.CodeFlix.Catalog.UnitTests.Application.Common;
 using System;
 using System.Collections.Generic;
 using Xunit;
@@ -16,34 +14,9 @@ public class ListCategoriesTestFixtureCollection : ICollectionFixture<ListCatego
 
 }
 
-public class ListCategoriesTestFixture : BaseFixture
+public class ListCategoriesTestFixture : CategoryUseCasesBaseFixture
 {
-    public Mock<ICategoryRepository> GetRepositoryMock() => new();
 
-    public string GetValidCategoryName()
-    {
-        var categoryName = "";
-
-        while (categoryName.Length < 3)
-            categoryName = Faker.Commerce.Categories(1)[0];
-
-        if (categoryName.Length > 255)
-            categoryName = categoryName[..255];
-
-        return categoryName;
-    }
-
-    public string GetValidCategoryDescription()
-    {
-        var categoryDescription = Faker.Commerce.ProductDescription();
-
-
-        if (categoryDescription.Length > 10_000)
-            categoryDescription = categoryDescription[..10_000];
-
-
-        return categoryDescription;
-    }
 
     public List<Category> GetExampleCategoriesList(int length = 10)
     {
@@ -68,13 +41,4 @@ public class ListCategoriesTestFixture : BaseFixture
                  dir: random.Next(0, 10) > 5 ? SearchOrder.ASC : SearchOrder.DESC
              );
     }
-
-
-    public bool GetRandomBoolean()
-        => (new Random().NextDouble()) < 0.5;
-
-    public Category GetExampleCategory()
-        => new(GetValidCategoryName(), GetValidCategoryDescription(), GetRandomBoolean());
-
-
 }
