+++ tests/FC.CodeFlix.Catalog.UnitTests/Application/UpdateCategory/UpdateCategoryTestFixture.cs
@@ -1,9 +1,5 @@
-﻿using FC.CodeFlix.Catalog.Application.Interfaces;
-using FC.CodeFlix.Catalog.Application.UseCases.Category.UpdateCategory;
-using FC.CodeFlix.Catalog.Domain.Entity;
-using FC.CodeFlix.Catalog.Domain.Repository;
-using FC.CodeFlix.Catalog.UnitTests.Common;
-using Moq;
+﻿using FC.CodeFlix.Catalog.Application.UseCases.Category.UpdateCategory;
+using FC.CodeFlix.Catalog.UnitTests.Application.Common;
 using System;
 using Xunit;
 
@@ -16,39 +12,9 @@ public class UpdateCategoryTestFixtureCollection : ICollectionFixture<UpdateCate
 
 }
 
-public class UpdateCategoryTestFixture : BaseFixture
+public class UpdateCategoryTestFixture : CategoryUseCasesBaseFixture
 {
-    public string GetValidCategoryName()
-    {
-        var categoryName = "";
-
-        while (categoryName.Length < 3)
-            categoryName = Faker.Commerce.Categories(1)[0];
-
-        if (categoryName.Length > 255)
-            categoryName = categoryName[..255];
-
-        return categoryName;
-    }
-
-    public string GetValidCategoryDescription()
-    {
-        var categoryDescription = Faker.Commerce.ProductDescription();
-
-
-        if (categoryDescription.Length > 10_000)
-            categoryDescription = categoryDescription[..10_000];
-
-
-        return categoryDescription;
-    }
-
-    public bool GetRandomBoolean()
-        => (new Random().NextDouble()) < 0.5;
-
-    public Category GetExampleCategory()
-        => new(GetValidCategoryName(), GetValidCategoryDescription(), GetRandomBoolean());
-
+ 
     public UpdateCategoryInput GetInvalidInputShortName()
     {
         var invalidInputShortName = GetValidInput();
@@ -75,16 +41,10 @@ public class UpdateCategoryTestFixture : BaseFixture
         return invalidInputTooLongDescription;
     }
 
-
     public UpdateCategoryInput GetValidInput(Guid? id = null)
      => new(id ?? Guid.NewGuid(),
             GetValidCategoryName(),
             GetValidCategoryDescription(),
             GetRandomBoolean());
 
-    
-
-    public Mock<ICategoryRepository> GetRepositoryMock() => new();
-
-    public Mock<IUnitOfWork> GetUnitOfWorkMock() => new();
 }
